Funcao policy

+++ app/Http/Controllers/FuncaoController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Models\Funcao;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class FuncaoController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      *
@@ -14,37 +16,26 @@ class FuncaoController extends Controller
      */
     public function index()
     {
+        if (! Gate::allows('index-funcao')) {
+            abort(403);
+        }
         return Funcao::all();
     }
 
 
-        /**
-     * Display a listing of the resource to a front select.
-     *
-     * @return \Illuminate\Http\Response
-    */
-    public function front_select()
-    {
-        return Funcao::all(['id as value','descricao as text']);
-    }
-
     /**
-     * Store a newly created resource in storage.
+     * Display a listing of the resource with all relationships.
      *
-     * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function index_full()
     {
-        $request->validate(
-            rules: [
-                'descricao' => "required",
-            ]
-        );
-        return Funcao::create(
-            $request->all()
-        );
+        if (! Gate::allows('index-funcao')) {
+            abort(403);
+        }
+        return Funcao::all()->load('perfil');
     }
+
 
     /**
      * Display the specified resource.
@@ -54,8 +45,54 @@ class FuncaoController extends Controller
      */
     public function show(Funcao $funcao)
     {
+        if (! Gate::allows('show-funcao', $funcao)) {
+            abort(403);
+        }
         return $funcao;
     }
+
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Models\Funcao  $funcao
+     * @return \Illuminate\Http\Response
+     */
+    public function show_full(Funcao $funcao)
+    {
+        if (! Gate::allows('show-funcao', $funcao)) {
+            abort(403);
+        }
+        return $funcao->load('perfil');
+    }
+
+
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
+        if (! Gate::allows('create-funcao')) {
+            abort(403);
+        }
+        $request->validate(
+            rules: [
+                'perfil_id' => "required",
+                'descricao' => "required",
+                'centro_custo' => "required",
+                'data_inicio' => "required",
+                'data_fim' => "required",
+            ]
+        );
+        return Funcao::create(
+            $request->all()
+        );
+    }
+
 
     /**
      * Update the specified resource in storage.
@@ -66,9 +103,16 @@ class FuncaoController extends Controller
      */
     public function update(Request $request, Funcao $funcao)
     {
+        if (! Gate::allows('update-funcao', $funcao)) {
+            abort(403);
+        }
         $request->validate(
             rules: [
+                'perfil_id' => "required",
                 'descricao' => "required",
+                'centro_custo' => "required",
+                'data_inicio' => "required",
+                'data_fim' => "required",
             ]
         );
         $funcao->update(
@@ -76,6 +120,7 @@ class FuncaoController extends Controller
         );
         return $funcao;
     }
+
 
     /**
      * Remove the specified resource from storage.
@@ -85,6 +130,20 @@ class FuncaoController extends Controller
      */
     public function destroy(Funcao $funcao)
     {
+        if (! Gate::allows('delete-funcao', $funcao)) {
+            abort(403);
+        }
         return $funcao->delete();
+    }
+
+
+    /**
+     * Display a listing of the resource to a front select.
+     *
+     * @return \Illuminate\Http\Response
+    */
+    public function front_select()
+    {
+        return Funcao::all(['id as value','descricao as text']);
     }
 }
